+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Added
 
 * Support for resolving constant and function-like #define macros.
+* Support for Houdini installation path ($HFS) to resolve standard include files.
 
 ## [0.3.0] - 2026-03-30